pyproject.toml
[tool.poetry]
name = "monzo-crawler"
version = "0.1.0"
description = "A simple concurrent python web crawler with maximum depth limit"
authors = ["AbishekKJ <abishek.kj@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.28.2"
beautifulsoup4 = "^4.11.2"
json-log-formatter = "1.0"
pyyaml = "^6.0"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^7.2.2"
pytest-mock = "^3.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
webcrawler = "crawler.__main__:main"